refactor(challenge): extract runJsSubmission to @superteam-lms/challenge-executor
